+++ src/main/java/fr/mkadia/mkadiaapi/repositories/RoleRepository.java
@@ -9,8 +9,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-public interface RoleRepository extends JpaRepository<Role,Long> {
-    List<Role> findAllByIdIn(List<Long> IDs);
+public interface RoleRepository extends JpaRepository<Role,Integer> {
+    List<Role> findAllByIdIn(List<Integer> IDs);
     Optional<Role> findFirstByLabelContainingOrderByLabelAsc(String label);
     Set<Role> findAllByIsDefaultTrue();
